requirements.txt
python-dotenv==1.0.1
pyTelegramBotAPI==4.16.1


